refactor(threads): useThreadMessaging 外圈 interrupt/review 拆出
- 4600→4025 行：threadMessagingInterrupt 434 + threadMessagingReview 377
- 外圈先行断自递归环；retryCodexSendAfterThreadRefresh 未被引用无需随迁（核实）
- return 对象 40 键与 HEAD diff 为空；参数对象每渲染内联新建
- baseline 重生成 + 新文件登记豁免
验证：tsc 触及文件零 error；useThreadMessaging 4 失败与基线逐一同名；姊妹 5 测试文件全绿
（子代理 LargeGiraffe 执行，主代理复核验证）
