+++ src/features/threads/hooks/threadMessagingReview.ts
@@ -1,0 +1,377 @@
+import { useCallback } from "react";
+import type { DebugEntry, ReviewTarget, WorkspaceInfo } from "../../../types";
+import {
+  startReview as startReviewService,
+} from "../../../services/tauri";
+import type { AutoSessionMetadata } from "../../../services/tauri";
+import {
+  extractRpcErrorMessage,
+  parseReviewTarget,
+} from "../utils/threadNormalize";
+import {
+  buildReviewCommandText,
+  isInvalidReviewThreadIdError,
+} from "./threadMessagingHelpers";
+import { useReviewPrompt } from "./useReviewPrompt";
+
+type ThreadEngine =
+  | "claude"
+  | "codex"
+  | "gemini"
+  | "grok"
+  | "kimi"
+  | "opencode"
+  | "pi"
+  | "dsh"
+  | "qoder";
+
+type ReviewSendMessageOptions = {
+  skipPromptExpansion?: boolean;
+  autoSession?: AutoSessionMetadata | null;
+};
+
+type UseThreadMessagingReviewOptions = {
+  activeEngine: ThreadEngine;
+  activeWorkspace: WorkspaceInfo | null;
+  activeThreadId: string | null;
+  ensureThreadForActiveWorkspace: () => Promise<string | null>;
+  ensureThreadForWorkspace: (workspaceId: string) => Promise<string | null>;
+  isThreadIdCompatibleWithEngine: (
+    engine: ThreadEngine,
+    threadId: string,
+  ) => boolean;
+  markProcessing: (threadId: string, isProcessing: boolean) => void;
+  markReviewing: (threadId: string, isReviewing: boolean) => void;
+  onDebug?: (entry: DebugEntry) => void;
+  pushThreadErrorMessage: (
+    workspaceId: string,
+    threadId: string,
+    message: string,
+  ) => void;
+  resolveThreadEngine: (workspaceId: string, threadId: string) => ThreadEngine;
+  safeMessageActivity: () => void;
+  sendMessageToThread: (
+    workspace: WorkspaceInfo,
+    threadId: string,
+    text: string,
+    images?: string[],
+    options?: ReviewSendMessageOptions,
+  ) => Promise<unknown>;
+  setActiveTurnId: (threadId: string, turnId: string | null) => void;
+  startThreadForWorkspace: (
+    workspaceId: string,
+    options?: {
+      activate?: boolean;
+      engine?: ThreadEngine;
+      autoSession?: AutoSessionMetadata | null;
+    },
+  ) => Promise<string | null>;
+};
+
+export function useThreadMessagingReview({
+  activeEngine,
+  activeWorkspace,
+  activeThreadId,
+  ensureThreadForActiveWorkspace,
+  ensureThreadForWorkspace,
+  isThreadIdCompatibleWithEngine,
+  markProcessing,
+  markReviewing,
+  onDebug,
+  pushThreadErrorMessage,
+  resolveThreadEngine,
+  safeMessageActivity,
+  sendMessageToThread,
+  setActiveTurnId,
+  startThreadForWorkspace,
+}: UseThreadMessagingReviewOptions) {
+  const startReviewTarget = useCallback(
+    async (
+      target: ReviewTarget,
+      workspaceIdOverride?: string,
+    ): Promise<boolean> => {
+      const workspaceId = workspaceIdOverride ?? activeWorkspace?.id ?? null;
+      if (!workspaceId) {
+        return false;
+      }
+      let threadId = workspaceIdOverride
+        ? await ensureThreadForWorkspace(workspaceId)
+        : await ensureThreadForActiveWorkspace();
+      if (!threadId) {
+        return false;
+      }
+      const reviewExecutionEngine: "claude" | "codex" =
+        activeEngine === "claude" ? "claude" : "codex";
+      const threadEngine = resolveThreadEngine(workspaceId, threadId);
+      const reviewAutoSession: AutoSessionMetadata = {
+        sessionPurpose: "review-fallback",
+        visibility: "system-auto",
+        ownerFeature: "review",
+        autoArchive: false,
+        createdBy: "system",
+      };
+      const threadIdCompatible = isThreadIdCompatibleWithEngine(
+        reviewExecutionEngine,
+        threadId,
+      );
+      if (threadEngine !== reviewExecutionEngine || !threadIdCompatible) {
+        onDebug?.({
+          id: `${Date.now()}-client-review-thread-rebind`,
+          timestamp: Date.now(),
+          source: "client",
+          label: "review/thread rebind",
+          payload: {
+            workspaceId,
+            originalThreadId: threadId,
+            originalThreadEngine: threadEngine,
+            threadIdCompatible,
+            targetEngine: reviewExecutionEngine,
+          },
+        });
+        const reviewThreadId = await startThreadForWorkspace(workspaceId, {
+          activate: workspaceId === activeWorkspace?.id,
+          engine: reviewExecutionEngine,
+          autoSession: reviewAutoSession,
+        });
+        if (!reviewThreadId) {
+          return false;
+        }
+        threadId = reviewThreadId;
+      }
+
+      if (reviewExecutionEngine === "claude") {
+        const reviewWorkspace =
+          activeWorkspace && activeWorkspace.id === workspaceId
+            ? activeWorkspace
+            : null;
+        if (!reviewWorkspace) {
+          return false;
+        }
+        const reviewCommand = buildReviewCommandText(target);
+        onDebug?.({
+          id: `${Date.now()}-client-review-start`,
+          timestamp: Date.now(),
+          source: "client",
+          label: "review/start (cli command)",
+          payload: {
+            workspaceId,
+            threadId,
+            target,
+            command: reviewCommand,
+            engine: "claude",
+          },
+        });
+        await sendMessageToThread(
+          reviewWorkspace,
+          threadId,
+          reviewCommand,
+          [],
+          {
+            skipPromptExpansion: true,
+            autoSession: reviewAutoSession,
+          },
+        );
+        return true;
+      }
+
+      markProcessing(threadId, true);
+      markReviewing(threadId, true);
+      safeMessageActivity();
+      let reviewThreadId = threadId;
+      onDebug?.({
+        id: `${Date.now()}-client-review-start`,
+        timestamp: Date.now(),
+        source: "client",
+        label: "review/start",
+        payload: {
+          workspaceId,
+          threadId,
+          target,
+        },
+      });
+      try {
+        const runStartReview = async (
+          targetThreadId: string,
+          label:
+            | "review/start response"
+            | "review/start retry response" = "review/start response",
+        ) => {
+          const response = await startReviewService(
+            workspaceId,
+            targetThreadId,
+            target,
+            "inline",
+          );
+          onDebug?.({
+            id: `${Date.now()}-server-review-start`,
+            timestamp: Date.now(),
+            source: "server",
+            label,
+            payload: response,
+          });
+          return response;
+        };
+
+        let response = await runStartReview(reviewThreadId);
+        let rpcError = extractRpcErrorMessage(response);
+
+        if (rpcError && isInvalidReviewThreadIdError(rpcError)) {
+          const fallbackThreadId = await startThreadForWorkspace(workspaceId, {
+            activate: workspaceId === activeWorkspace?.id,
+            engine: "codex",
+            autoSession: reviewAutoSession,
+          });
+          if (fallbackThreadId && fallbackThreadId !== reviewThreadId) {
+            onDebug?.({
+              id: `${Date.now()}-client-review-thread-retry`,
+              timestamp: Date.now(),
+              source: "client",
+              label: "review/thread retry",
+              payload: {
+                workspaceId,
+                originalThreadId: reviewThreadId,
+                fallbackThreadId,
+                reason: rpcError,
+              },
+            });
+            markProcessing(reviewThreadId, false);
+            markReviewing(reviewThreadId, false);
+            reviewThreadId = fallbackThreadId;
+            markProcessing(reviewThreadId, true);
+            markReviewing(reviewThreadId, true);
+            response = await runStartReview(
+              reviewThreadId,
+              "review/start retry response",
+            );
+            rpcError = extractRpcErrorMessage(response);
+          }
+        }
+        if (rpcError) {
+          markProcessing(reviewThreadId, false);
+          markReviewing(reviewThreadId, false);
+          setActiveTurnId(reviewThreadId, null);
+          pushThreadErrorMessage(
+            workspaceId,
+            reviewThreadId,
+            `Review failed to start: ${rpcError}`,
+          );
+          safeMessageActivity();
+          return false;
+        }
+        return true;
+      } catch (error) {
+        markProcessing(reviewThreadId, false);
+        markReviewing(reviewThreadId, false);
+        onDebug?.({
+          id: `${Date.now()}-client-review-start-error`,
+          timestamp: Date.now(),
+          source: "error",
+          label: "review/start error",
+          payload: error instanceof Error ? error.message : String(error),
+        });
+        pushThreadErrorMessage(
+          workspaceId,
+          reviewThreadId,
+          error instanceof Error ? error.message : String(error),
+        );
+        safeMessageActivity();
+        return false;
+      }
+    },
+    [
+      activeEngine,
+      activeWorkspace,
+      ensureThreadForActiveWorkspace,
+      ensureThreadForWorkspace,
+      isThreadIdCompatibleWithEngine,
+      markProcessing,
+      markReviewing,
+      onDebug,
+      pushThreadErrorMessage,
+      resolveThreadEngine,
+      safeMessageActivity,
+      sendMessageToThread,
+      setActiveTurnId,
+      startThreadForWorkspace,
+    ],
+  );
+
+  const {
+    reviewPrompt,
+    openReviewPrompt,
+    closeReviewPrompt,
+    showPresetStep,
+    choosePreset,
+    highlightedPresetIndex,
+    setHighlightedPresetIndex,
+    highlightedBranchIndex,
+    setHighlightedBranchIndex,
+    highlightedCommitIndex,
+    setHighlightedCommitIndex,
+    handleReviewPromptKeyDown,
+    confirmBranch,
+    selectBranch,
+    selectBranchAtIndex,
+    selectCommit,
+    selectCommitAtIndex,
+    confirmCommit,
+    updateCustomInstructions,
+    confirmCustom,
+  } = useReviewPrompt({
+    activeWorkspace,
+    activeThreadId,
+    onDebug,
+    startReviewTarget,
+  });
+
+  const startReview = useCallback(
+    async (text: string) => {
+      if (!activeWorkspace || !text.trim()) {
+        return;
+      }
+      const trimmed = text.trim();
+      if (!trimmed.startsWith("/")) {
+        return;
+      }
+      const commandToken =
+        trimmed.slice(1).split(/\s+/, 1)[0]?.toLowerCase() ?? "";
+      if (commandToken !== "review") {
+        return;
+      }
+      const rest = trimmed.slice(commandToken.length + 1).trim();
+      if (!rest) {
+        openReviewPrompt();
+        return;
+      }
+
+      const target = parseReviewTarget(trimmed);
+      await startReviewTarget(target);
+    },
+    [activeWorkspace, openReviewPrompt, startReviewTarget],
+  );
+
+  return {
+    reviewPrompt,
+    openReviewPrompt,
+    closeReviewPrompt,
+    showPresetStep,
+    choosePreset,
+    highlightedPresetIndex,
+    setHighlightedPresetIndex,
+    highlightedBranchIndex,
+    setHighlightedBranchIndex,
+    highlightedCommitIndex,
+    setHighlightedCommitIndex,
+    handleReviewPromptKeyDown,
+    confirmBranch,
+    selectBranch,
+    selectBranchAtIndex,
+    selectCommit,
+    selectCommitAtIndex,
+    confirmCommit,
+    updateCustomInstructions,
+    confirmCustom,
+    startReviewTarget,
+    startReview,
+  };
+}
