+++ src/features/threads/hooks/useThreadMessaging.ts
@@ -4,7 +4,6 @@ import { useOrphanTurnWatchdog } from "./useOrphanTurnWatchdog";
 import {
   workspaceScopedDelete,
   workspaceScopedHas,
-  workspaceScopedSet,
 } from "./workspaceScopedMap";
 import type { Dispatch, MutableRefObject } from "react";
 import { useTranslation } from "react-i18next";
@@ -16,7 +15,6 @@ import type {
   ThreadTokenUsage,
   CustomPromptOption,
   DebugEntry,
-  ReviewTarget,
   WorkspaceInfo,
   BrowserContextSendAttachment,
   IntentCanvasContextSendAttachment,
@@ -33,11 +31,7 @@ import {
 import { emitMessagesForcePinBottom } from "../../../live-canvas/liveCanvasControls";
 import {
   sendUserMessage as sendUserMessageService,
-  startReview as startReviewService,
-  interruptTurn as interruptTurnService,
-  engineInterruptTurn as engineInterruptTurnService,
   engineSendMessage as engineSendMessageService,
-  engineInterrupt as engineInterruptService,
   listGeminiSessions as listGeminiSessionsService,
   listGrokSessions as listGrokSessionsService,
   listKimiSessions as listKimiSessionsService,
@@ -52,13 +46,10 @@ import {
 } from "../../shared-session/runtime/sendSharedSessionTurnV2";
 import { subscribeSharedSessionAttemptSettlements } from "../../shared-session/runtime/reattachSharedSessionAttempt";
 import { isSharedV2SendEnabled } from "../../shared-session/runtime/sharedV2SendFlag";
-import { sharedSessionV2InterruptTurn as sharedSessionV2InterruptTurnService } from "../../shared-session/services/sharedSessions";
 import {
-  dispatchSharedSendEvent,
   getSharedSendActiveAttemptId,
   getSharedSendState,
   releaseSharedSendAdmission,
-  setSharedSendActiveAttempt,
   tryAcquireSharedSend,
 } from "../../shared-session/runtime/sharedSendStateStore";
 import {
@@ -226,10 +217,8 @@ import { expandCustomPromptText } from "../../../utils/customPrompts";
 import {
   asString,
   extractRpcErrorMessage,
-  parseReviewTarget,
 } from "../utils/threadNormalize";
 import type { ThreadAction, ThreadState } from "./useThreadsReducer";
-import { useReviewPrompt } from "./useReviewPrompt";
 import { pushErrorToast } from "../../../services/toasts";
 import { pushThreadFailureRuntimeNotice } from "../../../services/globalRuntimeNotices";
 import { resolveAgentIconForAgent } from "../../../utils/agentIcons";
@@ -258,15 +247,12 @@ import {
   type SessionSpecLinkContext,
 } from "./threadMessagingSpecRoot";
 import {
-  buildReviewCommandText,
   extractSessionIdFromEngineSendResponse,
   resolveDshModelForSend,
   resolveDshSendFallbackCatalogId,
   isCodexMissingThreadBindingError,
-  isInvalidReviewThreadIdError,
   isLikelyForeignModelForGemini,
   isRecoverableCodexThreadBindingError,
-  isUnknownEngineInterruptTurnMethodError,
   classifyTurnStartReasonCode,
   mapNetworkErrorToUserMessage,
   normalizeAccessMode,
@@ -286,6 +272,8 @@ import {
 } from "../utils/stabilityDiagnostics";
 import { useThreadMessagingSessionTooling } from "./useThreadMessagingSessionTooling";
 import { useThreadMessagingThreadResolution } from "./useThreadMessagingThreadResolution";
+import { useThreadInterruptTurn } from "./threadMessagingInterrupt";
+import { useThreadMessagingReview } from "./threadMessagingReview";
 import {
   createOptimisticGeneratedImageProcessingItem,
   extractOptimisticGeneratedImagePrompt,
@@ -294,8 +282,6 @@ import {
   resolveCodexAcceptedTurnFact,
   shouldDeferCodexActivityUntilTurnAccepted,
 } from "../utils/codexConversationLiveness";
-import { drainLiveAssistantTextTail } from "../utils/liveAssistantTextChannel";
-import { drainLiveItemDeltaTail } from "../utils/liveItemDeltaChannel";
 import { formatBrowserContextPromptOnce } from "../../browser-agent";
 import {
   buildLocalizedMemoryScoutPreviewText,
@@ -364,10 +350,6 @@ type SendMessageToThreadFn = (
   images?: string[],
   options?: SendMessageOptions,
 ) => Promise<ThreadMessageDispatchResult>;
-
-type InterruptTurnOptions = {
-  reason?: "user-stop" | "queue-fusion" | "plan-handoff";
-};
 
 type HandleFusionStalledOptions = {
   message?: string | null;
@@ -3898,565 +3880,22 @@ export function useThreadMessaging({
     ],
   );
 
-  const interruptTurn = useCallback(
-    async (options?: InterruptTurnOptions) => {
-      if (!activeWorkspace || !activeThreadId) {
-        return;
-      }
-      const reason = options?.reason ?? "user-stop";
-      if (activeWorkspace && activeThreadId) {
-        cancelSharedProviderRetry(activeWorkspace.id, activeThreadId, "stopped");
-      }
-      const activeThreadKind = resolveThreadKind(
-        activeWorkspace.id,
-        activeThreadId,
-      );
-      const usesSharedV2Control =
-        activeThreadKind === "shared" && isSharedV2SendEnabled();
-      const sharedAttemptId = usesSharedV2Control
-        ? getSharedSendActiveAttemptId(activeWorkspace.id, activeThreadId)
-        : null;
-      const activeTurnId = activeTurnIdByThread[activeThreadId] ?? null;
-      const activeThreadIsProcessing =
-        threadStatusById[activeThreadId]?.isProcessing ?? false;
-      if (!activeTurnId && !activeThreadIsProcessing && !usesSharedV2Control) {
-        onDebug?.({
-          id: `${Date.now()}-client-turn-interrupt-skipped`,
-          timestamp: Date.now(),
-          source: "client",
-          label: "turn/interrupt skipped",
-          payload: {
-            workspaceId: activeWorkspace.id,
-            threadId: activeThreadId,
-            reason,
-            cause: "no-active-or-processing-turn",
-          },
-        });
-        return;
-      }
-      if (usesSharedV2Control && !sharedAttemptId) {
-        const sharedSendState = getSharedSendState(
-          activeWorkspace.id,
-          activeThreadId,
-        ).state;
-        if (
-          sharedSendState === "idle" &&
-          (activeTurnId || activeThreadIsProcessing)
-        ) {
-          // canonical commit 已把 Shared send state 收口并释放 Attempt；此时只剩
-          // frontend lifecycle residue。它不再需要、也不允许触发 Runtime interrupt。
-          markProcessing(activeThreadId, false);
-          setActiveTurnId(activeThreadId, null);
-          onDebug?.({
-            id: `${Date.now()}-client-shared-turn-residue-converged`,
-            timestamp: Date.now(),
-            source: "client",
-            label: "shared-session/turn residue converged",
-            payload: {
-              workspaceId: activeWorkspace.id,
-              threadId: activeThreadId,
-              reason,
-              sharedSendState,
-            },
-          });
-          return;
-        }
-        onDebug?.({
-          id: `${Date.now()}-client-turn-interrupt-skipped`,
-          timestamp: Date.now(),
-          source: "client",
-          label: "turn/interrupt skipped",
-          payload: {
-            workspaceId: activeWorkspace.id,
-            threadId: activeThreadId,
-            reason,
-            cause: "shared-attempt-owner-missing",
-            sharedSendState,
-          },
-        });
-        return;
-      }
-      if (sharedAttemptId) {
-        try {
-          const interruptResult = await sharedSessionV2InterruptTurnService(
-            activeWorkspace.id,
-            activeThreadId,
-            sharedAttemptId,
-          );
-          if (interruptResult.status === "terminal-committed") {
-            dispatchSharedSendEvent(activeWorkspace.id, activeThreadId, {
-              type: "terminalCommitted",
-            });
-            setSharedSendActiveAttempt(
-              activeWorkspace.id,
-              activeThreadId,
-              null,
-            );
-            markProcessing(activeThreadId, false);
-            setActiveTurnId(activeThreadId, null);
-            return;
-          }
-        } catch (error) {
-          onDebug?.({
-            id: `${Date.now()}-client-turn-interrupt-error`,
-            timestamp: Date.now(),
-            source: "error",
-            label: "turn/interrupt error",
-            payload: error instanceof Error ? error.message : String(error),
-          });
-          return;
-        }
-      }
-      const turnId = activeTurnId ?? "pending";
-      const shouldGuardInterruptedThread = reason !== "queue-fusion";
-      // A4 live-text 外部化：中断前把通道里「尚未落 reducer 的尾段」灌回 items，
-      // 否则中断后该行会从通道全量文本回退到壳首段。hasCustomName: true 表示
-      // 灌回不参与线程自动命名。
-      const liveTextTail = drainLiveAssistantTextTail(activeThreadId);
-      if (liveTextTail) {
-        dispatch({
-          type: "appendAgentDelta",
-          workspaceId: activeWorkspace.id,
-          threadId: activeThreadId,
-          itemId: liveTextTail.itemId,
-          delta: liveTextTail.tailDelta,
-          hasCustomName: true,
-        });
-      }
-      // A4 二期：中断同样把 reasoning/toolOutput 通道里「尚未落 reducer 的尾段」
-      // 灌回 items（flag 关时通道为空、天然 no-op），否则中断后这些行会回退到
-      // 建壳首段。
-      for (const tail of drainLiveItemDeltaTail(activeThreadId)) {
-        if (tail.lane === "reasoningContent") {
-          dispatch({
-            type: "appendReasoningContent",
-            threadId: activeThreadId,
-            itemId: tail.itemId,
-            delta: tail.text,
-          });
-        } else if (tail.lane === "reasoningSummary") {
-          dispatch({
-            type: "appendReasoningSummary",
-            threadId: activeThreadId,
-            itemId: tail.itemId,
-            delta: tail.text,
-          });
-        } else {
-          dispatch({
-            type: "appendToolOutput",
-            threadId: activeThreadId,
-            itemId: tail.itemId,
-            delta: tail.text,
-          });
-        }
-      }
-      // Queue fusion immediately starts a successor turn on the same curtain; a
-      // long-lived interrupted guard would drop that successor's realtime output.
-      if (shouldGuardInterruptedThread) {
-        workspaceScopedSet(
-          interruptedThreadsRef.current,
-          activeWorkspace.id,
-          activeThreadId,
-          true,
-        );
-      }
-      markProcessing(activeThreadId, false);
-      setActiveTurnId(activeThreadId, null);
-      const interruptNotice =
-        reason === "queue-fusion"
-          ? t("threads.sessionStoppedForFusion")
-          : reason === "plan-handoff"
-            ? null
-            : t("threads.sessionStopped");
-      if (interruptNotice) {
-        dispatch({
-          type: "addAssistantMessage",
-          threadId: activeThreadId,
-          text: interruptNotice,
-        });
-      }
-      if (!activeTurnId && shouldGuardInterruptedThread) {
-        workspaceScopedSet(
-          pendingInterruptsRef.current,
-          activeWorkspace.id,
-          activeThreadId,
-          true,
-        );
-      }
-
-      // Determine whether this thread is backed by a local CLI session.
-      const resolvedThreadEngine = resolveThreadEngine(
-        activeWorkspace.id,
-        activeThreadId,
-      );
-      const isCliManagedEngine = resolvedThreadEngine !== "codex";
-
-      onDebug?.({
-        id: `${Date.now()}-client-turn-interrupt`,
-        timestamp: Date.now(),
-        source: "client",
-        label: "turn/interrupt",
-        payload: {
-          workspaceId: activeWorkspace.id,
-          threadId: activeThreadId,
-          turnId,
-          queued: !activeTurnId,
-          engine: resolvedThreadEngine,
-          reason,
-        },
-      });
-      try {
-        const sharedProviderProfileId =
-          activeThreadKind === "shared"
-            ? (getSharedTargetState(activeWorkspace.id, activeThreadId)
-                .activeTurnTarget?.providerProfileId ?? null)
-            : null;
-        // Qoder Global/CN are two runtimes behind one engine id. Native Qoder
-        // threads must carry their persisted distribution binding when
-        // interrupting; omitting it intentionally resolves the legacy Global
-        // runtime in Rust.
-        const nativeQoderStoredProfileId =
-          activeThreadKind === "native" && resolvedThreadEngine === "qoder"
-            ? (getThreadProviderProfileId?.(
-                activeWorkspace.id,
-                activeThreadId,
-              ) ?? null)
-            : null;
-        const nativeQoderIdentity =
-          activeThreadKind === "native" &&
-          resolvedThreadEngine === "qoder" &&
-          activeThreadId.startsWith("qoder:")
-            ? parseQoderSessionIdentity(
-                activeThreadId,
-                nativeQoderStoredProfileId,
-              )
-            : null;
-        if (
-          activeThreadKind === "native" &&
-          resolvedThreadEngine === "qoder" &&
-          activeThreadId.startsWith("qoder:") &&
-          !nativeQoderIdentity
-        ) {
-          onDebug?.({
-            id: `${Date.now()}-client-qoder-interrupt-identity-rejected`,
-            timestamp: Date.now(),
-            source: "client",
-            label: "turn/interrupt Qoder identity rejected",
-            payload: { workspaceId: activeWorkspace.id, threadId: activeThreadId },
-          });
-          return;
-        }
-        const nativeQoderProviderProfileId =
-          resolvedThreadEngine === "qoder"
-            ? (nativeQoderIdentity?.providerProfileId ??
-              canonicalQoderProviderProfileId(nativeQoderStoredProfileId))
-            : null;
-        if (usesSharedV2Control) {
-          // Shared V2 已由 durable attempt owner 精确中断；禁止再走 mutable
-          // target / workspace-wide fallback 产生第二次 control side effect。
-          onDebug?.({
-            id: `${Date.now()}-server-turn-interrupt`,
-            timestamp: Date.now(),
-            source: "server",
-            label: "turn/interrupt response",
-            payload: { success: true },
-          });
-          return;
-        }
-        if (isCliManagedEngine) {
-          // Claude/OpenCode/Gemini: target only the current turn process.
-          // If turn id is not known yet, keep pending interrupt and let onTurnStarted
-          // execute a precise kill once the backend emits the real turn id.
-          if (activeTurnId) {
-            try {
-              if (activeThreadKind === "shared" || nativeQoderProviderProfileId) {
-                await engineInterruptTurnService(
-                  activeWorkspace.id,
-                  activeTurnId,
-                  resolvedThreadEngine,
-                  sharedProviderProfileId ?? nativeQoderProviderProfileId,
-                );
-              } else {
-                await engineInterruptTurnService(
-                  activeWorkspace.id,
-                  activeTurnId,
-                  resolvedThreadEngine,
-                );
-              }
-            } catch (error) {
-              if (
-                isUnknownEngineInterruptTurnMethodError(error) &&
-                resolvedThreadEngine !== "qoder"
-              ) {
-                // Compatibility fallback for stale daemon/runtime that doesn't
-                // implement engine_interrupt_turn yet.
-                await engineInterruptService(activeWorkspace.id);
-              } else {
-                // Qoder Global/CN 不能降级到 workspace-wide interrupt：旧 RPC
-                // 无法携带 distribution，可能误中断同 workspace 的另一套 runtime。
-                throw error;
-              }
-            }
-          }
-        } else {
-          // Codex: notify daemon via turn_interrupt RPC, plus engine_interrupt fallback.
-          // B.5：Shared Thread 按 active Turn 的 Execution Target provider 路由，
-          // 避免同 engine 双 Provider 并行时中断打到 default Provider 会话。
-          await Promise.allSettled([
-            interruptTurnService(
-              activeWorkspace.id,
-              activeThreadId,
-              turnId,
-              sharedProviderProfileId,
-            ),
-            engineInterruptService(activeWorkspace.id),
-          ]);
-        }
-        onDebug?.({
-          id: `${Date.now()}-server-turn-interrupt`,
-          timestamp: Date.now(),
-          source: "server",
-          label: "turn/interrupt response",
-          payload: { success: true },
-        });
-      } catch (error) {
-        onDebug?.({
-          id: `${Date.now()}-client-turn-interrupt-error`,
-          timestamp: Date.now(),
-          source: "error",
-          label: "turn/interrupt error",
-          payload: error instanceof Error ? error.message : String(error),
-        });
-      }
-    },
-    [
-      activeThreadId,
-      activeTurnIdByThread,
-      activeWorkspace,
-      dispatch,
-      interruptedThreadsRef,
-      markProcessing,
-      onDebug,
-      pendingInterruptsRef,
-      getThreadProviderProfileId,
-      resolveThreadEngine,
-      resolveThreadKind,
-      setActiveTurnId,
-      t,
-      threadStatusById,
-    ],
-  );
-
-  const startReviewTarget = useCallback(
-    async (
-      target: ReviewTarget,
-      workspaceIdOverride?: string,
-    ): Promise<boolean> => {
-      const workspaceId = workspaceIdOverride ?? activeWorkspace?.id ?? null;
-      if (!workspaceId) {
-        return false;
-      }
-      let threadId = workspaceIdOverride
-        ? await ensureThreadForWorkspace(workspaceId)
-        : await ensureThreadForActiveWorkspace();
-      if (!threadId) {
-        return false;
-      }
-      const reviewExecutionEngine: "claude" | "codex" =
-        activeEngine === "claude" ? "claude" : "codex";
-      const threadEngine = resolveThreadEngine(workspaceId, threadId);
-      const reviewAutoSession: AutoSessionMetadata = {
-        sessionPurpose: "review-fallback",
-        visibility: "system-auto",
-        ownerFeature: "review",
-        autoArchive: false,
-        createdBy: "system",
-      };
-      const threadIdCompatible = isThreadIdCompatibleWithEngine(
-        reviewExecutionEngine,
-        threadId,
-      );
-      if (threadEngine !== reviewExecutionEngine || !threadIdCompatible) {
-        onDebug?.({
-          id: `${Date.now()}-client-review-thread-rebind`,
-          timestamp: Date.now(),
-          source: "client",
-          label: "review/thread rebind",
-          payload: {
-            workspaceId,
-            originalThreadId: threadId,
-            originalThreadEngine: threadEngine,
-            threadIdCompatible,
-            targetEngine: reviewExecutionEngine,
-          },
-        });
-        const reviewThreadId = await startThreadForWorkspace(workspaceId, {
-          activate: workspaceId === activeWorkspace?.id,
-          engine: reviewExecutionEngine,
-          autoSession: reviewAutoSession,
-        });
-        if (!reviewThreadId) {
-          return false;
-        }
-        threadId = reviewThreadId;
-      }
-
-      if (reviewExecutionEngine === "claude") {
-        const reviewWorkspace =
-          activeWorkspace && activeWorkspace.id === workspaceId
-            ? activeWorkspace
-            : null;
-        if (!reviewWorkspace) {
-          return false;
-        }
-        const reviewCommand = buildReviewCommandText(target);
-        onDebug?.({
-          id: `${Date.now()}-client-review-start`,
-          timestamp: Date.now(),
-          source: "client",
-          label: "review/start (cli command)",
-          payload: {
-            workspaceId,
-            threadId,
-            target,
-            command: reviewCommand,
-            engine: "claude",
-          },
-        });
-        await sendMessageToThread(
-          reviewWorkspace,
-          threadId,
-          reviewCommand,
-          [],
-          {
-            skipPromptExpansion: true,
-            autoSession: reviewAutoSession,
-          },
-        );
-        return true;
-      }
-
-      markProcessing(threadId, true);
-      markReviewing(threadId, true);
-      safeMessageActivity();
-      let reviewThreadId = threadId;
-      onDebug?.({
-        id: `${Date.now()}-client-review-start`,
-        timestamp: Date.now(),
-        source: "client",
-        label: "review/start",
-        payload: {
-          workspaceId,
-          threadId,
-          target,
-        },
-      });
-      try {
-        const runStartReview = async (
-          targetThreadId: string,
-          label:
-            | "review/start response"
-            | "review/start retry response" = "review/start response",
-        ) => {
-          const response = await startReviewService(
-            workspaceId,
-            targetThreadId,
-            target,
-            "inline",
-          );
-          onDebug?.({
-            id: `${Date.now()}-server-review-start`,
-            timestamp: Date.now(),
-            source: "server",
-            label,
-            payload: response,
-          });
-          return response;
-        };
-
-        let response = await runStartReview(reviewThreadId);
-        let rpcError = extractRpcErrorMessage(response);
-
-        if (rpcError && isInvalidReviewThreadIdError(rpcError)) {
-          const fallbackThreadId = await startThreadForWorkspace(workspaceId, {
-            activate: workspaceId === activeWorkspace?.id,
-            engine: "codex",
-            autoSession: reviewAutoSession,
-          });
-          if (fallbackThreadId && fallbackThreadId !== reviewThreadId) {
-            onDebug?.({
-              id: `${Date.now()}-client-review-thread-retry`,
-              timestamp: Date.now(),
-              source: "client",
-              label: "review/thread retry",
-              payload: {
-                workspaceId,
-                originalThreadId: reviewThreadId,
-                fallbackThreadId,
-                reason: rpcError,
-              },
-            });
-            markProcessing(reviewThreadId, false);
-            markReviewing(reviewThreadId, false);
-            reviewThreadId = fallbackThreadId;
-            markProcessing(reviewThreadId, true);
-            markReviewing(reviewThreadId, true);
-            response = await runStartReview(
-              reviewThreadId,
-              "review/start retry response",
-            );
-            rpcError = extractRpcErrorMessage(response);
-          }
-        }
-        if (rpcError) {
-          markProcessing(reviewThreadId, false);
-          markReviewing(reviewThreadId, false);
-          setActiveTurnId(reviewThreadId, null);
-          pushThreadErrorMessage(
-            workspaceId,
-            reviewThreadId,
-            `Review failed to start: ${rpcError}`,
-          );
-          safeMessageActivity();
-          return false;
-        }
-        return true;
-      } catch (error) {
-        markProcessing(reviewThreadId, false);
-        markReviewing(reviewThreadId, false);
-        onDebug?.({
-          id: `${Date.now()}-client-review-start-error`,
-          timestamp: Date.now(),
-          source: "error",
-          label: "review/start error",
-          payload: error instanceof Error ? error.message : String(error),
-        });
-        pushThreadErrorMessage(
-          workspaceId,
-          reviewThreadId,
-          error instanceof Error ? error.message : String(error),
-        );
-        safeMessageActivity();
-        return false;
-      }
-    },
-    [
-      activeEngine,
-      activeWorkspace,
-      ensureThreadForActiveWorkspace,
-      ensureThreadForWorkspace,
-      isThreadIdCompatibleWithEngine,
-      markProcessing,
-      markReviewing,
-      onDebug,
-      pushThreadErrorMessage,
-      resolveThreadEngine,
-      safeMessageActivity,
-      sendMessageToThread,
-      setActiveTurnId,
-      startThreadForWorkspace,
-    ],
-  );
+  const interruptTurn = useThreadInterruptTurn({
+    activeThreadId,
+    activeTurnIdByThread,
+    activeWorkspace,
+    dispatch,
+    interruptedThreadsRef,
+    markProcessing,
+    onDebug,
+    pendingInterruptsRef,
+    getThreadProviderProfileId,
+    resolveThreadEngine,
+    resolveThreadKind,
+    setActiveTurnId,
+    t,
+    threadStatusById,
+  });
 
   const {
     reviewPrompt,
@@ -4479,38 +3918,24 @@ export function useThreadMessaging({
     confirmCommit,
     updateCustomInstructions,
     confirmCustom,
-  } = useReviewPrompt({
+    startReview,
+  } = useThreadMessagingReview({
+    activeEngine,
     activeWorkspace,
     activeThreadId,
+    ensureThreadForActiveWorkspace,
+    ensureThreadForWorkspace,
+    isThreadIdCompatibleWithEngine,
+    markProcessing,
+    markReviewing,
     onDebug,
-    startReviewTarget,
+    pushThreadErrorMessage,
+    resolveThreadEngine,
+    safeMessageActivity,
+    sendMessageToThread,
+    setActiveTurnId,
+    startThreadForWorkspace,
   });
-
-  const startReview = useCallback(
-    async (text: string) => {
-      if (!activeWorkspace || !text.trim()) {
-        return;
-      }
-      const trimmed = text.trim();
-      if (!trimmed.startsWith("/")) {
-        return;
-      }
-      const commandToken =
-        trimmed.slice(1).split(/\s+/, 1)[0]?.toLowerCase() ?? "";
-      if (commandToken !== "review") {
-        return;
-      }
-      const rest = trimmed.slice(commandToken.length + 1).trim();
-      if (!rest) {
-        openReviewPrompt();
-        return;
-      }
-
-      const target = parseReviewTarget(trimmed);
-      await startReviewTarget(target);
-    },
-    [activeWorkspace, openReviewPrompt, startReviewTarget],
-  );
 
   const {
     startCompact,
